pyproject.toml
[project]
name = "etl-nf-ia"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.2.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pymupdf>=1.25.1",
    "langchain-openai>=1.1.7",
    "sqlalchemy>=2.0.45",
    "psycopg2-binary>=2.9.11",
]

[dependency-groups]
dev = [
    "black",
    "isort",
    "mypy",
    "pytest>=8.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]
pythonpath = [
    ".",
]
